Core/DataStores: Updated db2 structures to 9.1.5
Port From (https://github.com/TrinityCore/TrinityCore/commit/d52c00a8fe79186013980af0e2caa3fd13c0c5fa)

+++ Source/Framework/Constants/GarrisonConst.cs
@@ -23,7 +23,7 @@ namespace Framework.Constants
         public const uint Alliance = 1;
     }
 
-    public enum GarrisonBuildingFlags : byte
+    public enum GarrisonBuildingFlags : int
     {
         NeedsPlan = 0x1
     }
@@ -42,7 +42,7 @@ namespace Framework.Constants
         Covenant = 123
     }
 
-    public enum GarrisonAbilityFlags : ushort
+    public enum GarrisonAbilityFlags : int
     {
         Trait = 0x01,
         CannotRoll = 0x02,
